replace rCore blk driver with svdriver
