+++ domain-tools/basic/src/io/mod.rs
@@ -1,2 +1,4 @@
 mod region;
 pub use region::SafeIORegion;
+mod construct;
+pub use construct::*;
